Updated formula
Updated the formula as per:
https://docs.google.com/spreadsheets/d/1KpMnCdzFHmfU0XDzUon5XviRis1MvlB5M6Y8fyIvcmo#gid=367847530

Changes of this new formula:
- Constant K (rate of decay) as been replaced by a minimum ranking in
the leaderboard of 60%. Wich is much more manageable.
- This also means the worst rank allowed points isn't ~25th anymore, but
increases with the amount of players. While the bonus points given to
top ranks for a larger leaderboard size still reduces drastically after
a while (see latural logarithm function).
- The maximum score/rank raises with the minimum rank permitted (meaning
the more players given a score, the more points given to the bests).
- The exponential decay curve for the score/rank is now a logit-looking
curve.
- No need for conditions  as invalid rank/leaderboard_size will now
return 0 rather than #INVALID_NUMBER.
- The amount of points given for top ranks has been very slightly
decreased in small leaderboards and slighlty increased in large
leaderboardas a side effect.

+formula